chore:() finish project configuration
__DESCRIPTION:__
The configuration of the project has been completed. It has been created
and split into modules to facilitate the maintenance and scalability of
the project. The configuration files have been created and the project
has been prepared for the development of the first features.

The environment variables are beoing loaded using infisical, we plan to
self-host the configuration files in the future or self-host the
infisical service through a docker container.

Also the project resources have been splited into modules withing the
entries folder, this will allow us to have a better organization of the
project resources and to facilitate the maintenance and scalability of
the project. All the resources are injected into the app module via
spread operator.
__CHANGELIST:__
-	modified:   apps/api/src/app/app.module.ts
-	new file:   apps/api/src/app/entries/index.ts
-	new file:   apps/api/src/app/entries/modules.ts
-	new file:   apps/api/src/app/entries/providers.ts
-	new file:   apps/api/src/config/apollo.config.ts
-	new file:   apps/api/src/config/config.ts
-	new file:   apps/api/src/config/index.ts
-	new file:   apps/api/src/config/namespaces/app.config.ts
-	new file:   apps/api/src/config/namespaces/database.config.ts
-	new file:   apps/api/src/config/typeorm.config.ts
-	modified:   apps/api/src/main.ts